+++ src/main/java/com/jm/jobseekerplatform/model/profiles/SeekerProfile.java
@@ -2,6 +2,7 @@ package com.jm.jobseekerplatform.model.profiles;
 
 import com.jm.jobseekerplatform.model.Portfolio;
 import com.jm.jobseekerplatform.model.Tag;
+import com.jm.jobseekerplatform.model.Vacancy;
 import org.hibernate.annotations.Type;
 
 import javax.persistence.*;
