+++ src/main/resources/templates/admin/admin_vacancies.html
@@ -1,8 +1,7 @@
 <!doctype html>
 <html xmlns:th="http://www.thymeleaf.org" lang="en"
       xmlns:layout="http://www.ultraq.net.nz/thymeleaf/layout"
-      layout:decorator="layouts/adminlayouts">
-
+      layout:decorator="layouts/adminlayouts" th:with="side='vac'">
 <head>
     <!--<link rel="stylesheet" type="text/css" href="https://cdn.datatables.net/1.10.19/css/jquery.dataTables.css">-->
     <link rel="stylesheet" type="text/css" href="https://cdn.datatables.net/1.10.19/css/dataTables.bootstrap4.min.css">
@@ -10,9 +9,7 @@
 </head>
 
 <body>
-
 <div class="container-fluid" layout:fragment="content">
-    <main role="main" class="col-md-9 ml-sm-auto col-lg-10 px-4" id="main_admin_page">
         <br/>
         <h2>Все вакансии</h2>
         <div class="table-responsive">
@@ -93,11 +90,10 @@
             </div>
 
         </div>
-    </main>
 </div>
 
 <div class="container-fluid" layout:fragment="content-script">
-    <script src="/js/admin_vacancies.js"></script>
+    <script src="/js/admin/admin_vacancies.js"></script>
     <script type="text/javascript" charset="utf8"
             src="https://cdn.datatables.net/1.10.19/js/jquery.dataTables.js"></script>
     <script src="https://cdn.datatables.net/1.10.19/js/dataTables.bootstrap4.min.js"></script>
