--- src/main/resources/templates/admin_vacancies.html
@@ -1,96 +1,0 @@
-<!doctype html>
-<html xmlns:th="http://www.thymeleaf.org" lang="en"
-      xmlns:layout="http://www.ultraq.net.nz/thymeleaf/layout"
-      layout:decorator="layouts/adminlayouts" th:with="side='vac'">
-<head>
-    <!--<link rel="stylesheet" type="text/css" href="https://cdn.datatables.net/1.10.19/css/jquery.dataTables.css">-->
-    <link rel="stylesheet" type="text/css" href="https://cdn.datatables.net/1.10.19/css/dataTables.bootstrap4.min.css">
-    <!--<link rel="stylesheet" type="text/css" href=https://cdn.datatables.net/1.10.19/css/dataTables.bootstrap.min.css>-->
-</head>
-<body>
-<div class="container-fluid" layout:fragment="content">
-    <br/>
-    <h2>Все вакансии</h2>
-    <div class="table-responsive">
-        <table id="vacancyTable" class="table table-striped table-sm table-bordered" cellspacing="0"
-               width="100%">
-            <thead>
-            <tr>
-                <th>Id</th>
-                <th>Headline</th>
-                <th>City</th>
-                <th>Min</th>
-                <th>Max</th>
-                <th>State</th>
-                <th>Chat</th>
-                <th>Edit</th>
-            </tr>
-            </thead>
-            <tbody>
-            <div id="editModal" class="modal fade" tabindex="-1">
-                <div class="modal-dialog modal-sm">
-                    <div class="modal-content">
-                        <div class="modal-header">
-                            <button class="close" id="modalclosebutton" data-dismiss="modal">x</button>
-                            <!--<h4 class="modal-title">Edit user </h4>-->
-                        </div>
-                        <div class="modal-body">
-                            <form id="modalform" method="post">
-                                <div class="form-group">
-                                    <label for="Id">Id</label>
-                                    <input type="text" class="form-control" id="Id" name="Id"
-                                           placeholder="Id"
-                                           readonly>
-                                </div>
-                                <div class="form-group">
-                                    <label for="Headline">Headline</label>
-                                    <input type="text" class="form-control" id="Headline" name="Headline"
-                                           placeholder="Headline" readonly>
-                                </div>
-                                <div class="form-group">
-                                    <label for="City">City</label>
-                                    <input type="text" class="form-control" id="City" name="City"
-                                           placeholder="City" readonly>
-                                </div>
-                                <div class="form-group">
-                                    <label for="State">State</label>
-                                    <select class="form-control" id="State">
-                                        <option value="ACCESS">Access</option>
-                                        <option value="NO_ACCESS">No Access</option>
-                                        <option value="BLOCK">Block</option>
-                                    </select>
-                                </div>
-                                <div class="modal-footer">
-                                    <input type="submit" class="btn btn-success btn-block" id="modalbutton"
-                                           value="Edit">
-                                </div>
-                            </form>
-                        </div>
-                    </div>
-
-                </div>
-            </div>
-            </tbody>
-            <tfoot>
-            <tr>
-                <th>Id</th>
-                <th>Headline</th>
-                <th>City</th>
-                <th>Min</th>
-                <th>Max</th>
-                <th>State</th>
-                <th>Chat</th>
-                <th>Edit</th>
-            </tr>
-            </tfoot>
-        </table>
-    </div>
-</div>
-<div class="container-fluid" layout:fragment="content-script">
-    <script src="/js/admin/admin_vacancies.js"></script>
-    <script type="text/javascript" charset="utf8"
-            src="https://cdn.datatables.net/1.10.19/js/jquery.dataTables.js"></script>
-    <script src="https://cdn.datatables.net/1.10.19/js/dataTables.bootstrap4.min.js"></script>
-</div>
-</body>
-</html>
